refactor(systems-grid): extract gamepad grid navigation into a part file
Move the gamepad/keyboard navigation slice of _SystemCardGridViewState
(_initializeGamepad, _cleanupGamepad, _navigateGrid, _navigateVirtual,
_ensureSelectedItemVisibleUniversal) verbatim into a new part file,
my_systems_grid/gamepad_grid_nav.dart, as `extension _GamepadGridNav`.

Pure source-split, no behaviour or public-API change. All state stays on
the host State (extensions can't declare fields); the 4 setState calls route
through a new host `rebuild()` bridge (State.setState is @protected and can't
be invoked from an extension). The memoization wrappers _buildVirtualGrid /
_calculateGridDimensions stay in the host (shared with build()).

Host 1496 -> 1300; new file 211 lines. Moved bodies are byte-identical to
main (token-stream verified). Local gate: dart format clean, flutter analyze
"No issues found", flutter test 207/207.
